requirements.txt
Flask==3.0.0
flask-cors==6.0.1
flasgger==0.9.7.1
python-dotenv==1.0.0
requests==2.32.3
